1202_2교시

+++ java_project/src/ch2_1017/CharacterEx1.java
@@ -8,16 +8,21 @@ public class CharacterEx1 {
 		//ch1 변수를 형변환, 명시적 형변환. 
 		// 캐스팅 연산자를 갑자기 사용하고 있음. 당황하지 마세요.
 		//7장 까지는 가볍게 들어보시고, 복습하면 됩니다. 
-		System.out.println((int)ch1);
+		
+		
+		System.out.println((int)ch1); // 65
 		System.out.println("=====================");
+		
 		char ch2 = 66;
 		System.out.println(ch2);
-		System.out.println(ch2);
+		System.out.println((int)ch2); // B
 		System.out.println("=====================");
+		
 		int i1 = 67;
 		System.out.println(i1);
 		System.out.println((char)i1);
 		System.out.println("=====================");
+		
 		char ch3 = '한';
 		char ch4 = '\uD55C';
